+++ obj/src/data/StringValueMap.js
@@ -63,8 +63,8 @@ var StringValueMap = (function () {
         }
     };
     StringValueMap.prototype.getAsObject = function (key) {
-        if (key === void 0) { key = null; }
-        if (key == null) {
+        if (key === void 0) { key = undefined; }
+        if (key === undefined) {
             var result = {};
             for (var key_1 in this) {
                 var value = this[key_1];
@@ -78,8 +78,8 @@ var StringValueMap = (function () {
         }
     };
     StringValueMap.prototype.setAsObject = function (key, value) {
-        if (value === void 0) { value = null; }
-        if (value == null) {
+        if (value === void 0) { value = undefined; }
+        if (value === undefined) {
             value = key;
             this.clear();
             var values = MapConverter_1.MapConverter.toMap(value);