+++ src/data/StringValueMap.ts
@@ -74,8 +74,8 @@ export class StringValueMap {
     	}
     }
         
-    public getAsObject(key: string = null): any {
-        if (key == null) {
+    public getAsObject(key: string = undefined): any {
+        if (key === undefined) {
             let result: any = {};
             for (let key in this) {
                 let value = this[key];
@@ -88,8 +88,8 @@ export class StringValueMap {
         }
     }
     
-    public setAsObject(key: any, value: any = null): void {
-        if (value == null) {
+    public setAsObject(key: any, value: any = undefined): void {
+        if (value === undefined) {
             value = key
             this.clear();
             let values = MapConverter.toMap(value);